backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
whitenoise==6.8.2
